requirements.txt
pyspark==3.5.1
presidio-analyzer>=2.2.353
presidio-anonymizer>=2.2.353
pydantic>=2.0.0
pytest>=7.0.0
pyyaml>=6.0.1
